+++ rust/crd/src/affinity.rs
@@ -90,9 +90,8 @@ mod test {
                                     ),
                                 ])),
                             }),
-                            namespace_selector: None,
-                            namespaces: None,
                             topology_key: "kubernetes.io/hostname".to_string(),
+                            ..PodAffinityTerm::default()
                         },
                         weight: 70,
                     },