+++ rust/crd/src/lib.rs
@@ -14,14 +14,16 @@ use product_config::{types::PropertyNameKind, ProductConfigManager};
 use s3logdir::S3LogDir;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_operator::builder::SecretFormat;
 use stackable_operator::product_config_utils::{
     transform_all_roles_to_config, validate_all_roles_and_groups_config,
     ValidatedRoleConfigByPropertyKind,
 };
 use stackable_operator::role_utils::EmptyRoleConfig;
 use stackable_operator::{
-    builder::{SecretOperatorVolumeSourceBuilder, VolumeBuilder},
+    builder::pod::volume::{
+        SecretFormat, SecretOperatorVolumeSourceBuilder, SecretOperatorVolumeSourceBuilderError,
+        VolumeBuilder,
+    },
     commons::{
         product_image_selection::{ProductImage, ResolvedProductImage},
         resources::{CpuLimits, MemoryLimits, Resources},
@@ -63,7 +65,7 @@ pub enum Error {
 
     #[snafu(display("failed to convert java heap config to unit [{unit}]"))]
     FailedToConvertJavaHeap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::memory::Error,
         unit: String,
     },
 
@@ -75,17 +77,17 @@ pub enum Error {
 
     #[snafu(display("failed to transform configs"))]
     ProductConfigTransform {
-        source: stackable_operator::product_config_utils::ConfigError,
+        source: stackable_operator::product_config_utils::Error,
     },
 
     #[snafu(display("invalid product config"))]
     InvalidProductConfig {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::product_config_utils::Error,
     },
 
     #[snafu(display("failed to build TLS certificate SecretClass Volume"))]
     TlsCertSecretClassVolumeBuild {
-        source: stackable_operator::builder::SecretOperatorVolumeSourceBuilderError,
+        source: SecretOperatorVolumeSourceBuilderError,
     },
 
     #[snafu(display("failed to build S3 credentials Volume"))]
@@ -1298,25 +1300,19 @@ mod tests {
                 mount_path: "/stackable/spark/driver-pod-templates".into(),
                 mount_propagation: None,
                 name: "driver-pod-template".into(),
-                read_only: None,
-                sub_path: None,
-                sub_path_expr: None,
+                ..VolumeMount::default()
             },
             VolumeMount {
                 mount_path: "/stackable/spark/executor-pod-templates".into(),
                 mount_propagation: None,
                 name: "executor-pod-template".into(),
-                read_only: None,
-                sub_path: None,
-                sub_path_expr: None,
+                ..VolumeMount::default()
             },
             VolumeMount {
                 mount_path: "/kerberos".into(),
                 mount_propagation: None,
                 name: "keytab".into(),
-                read_only: None,
-                sub_path: None,
-                sub_path_expr: None,
+                ..VolumeMount::default()
             },
         ];