+++ rust/operator-binary/src/history/history_controller.rs
@@ -4,7 +4,11 @@ use crate::Ctx;
 use product_config::{types::PropertyNameKind, writer::to_java_properties_string};
 use stackable_operator::kvp::{Label, Labels, ObjectLabels};
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder, VolumeBuilder},
+    builder::{
+        configmap::ConfigMapBuilder,
+        meta::ObjectMetaBuilder,
+        pod::{container::ContainerBuilder, volume::VolumeBuilder, PodBuilder},
+    },
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
     commons::product_image_selection::ResolvedProductImage,
     k8s_openapi::{
@@ -50,7 +54,7 @@ use std::collections::HashMap;
 use std::{collections::BTreeMap, sync::Arc};
 
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_operator::builder::resources::ResourceRequirementsBuilder;
+use stackable_operator::builder::pod::resources::ResourceRequirementsBuilder;
 use stackable_operator::k8s_openapi::DeepMerge;
 use stackable_operator::logging::controller::ReconcilerError;
 use strum::{EnumDiscriminants, IntoStaticStr};
@@ -65,36 +69,36 @@ pub enum Error {
     ObjectHasNoNamespace,
     #[snafu(display("invalid config map {name}"))]
     InvalidConfigMap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::configmap::Error,
         name: String,
     },
     #[snafu(display("invalid history container name"))]
     InvalidContainerName {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::pod::container::Error,
     },
     #[snafu(display("object is missing metadata to build owner reference"))]
     ObjectMissingMetadataForOwnerRef {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::meta::Error,
     },
     #[snafu(display("failed to update the history server deployment"))]
     ApplyDeployment {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to update history server config map"))]
     ApplyConfigMap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to update history server service"))]
     ApplyService {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to apply role ServiceAccount"))]
     ApplyServiceAccount {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to apply global RoleBinding"))]
     ApplyRoleBinding {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("product config validation failed"))]
     ProductConfigValidation {
@@ -114,11 +118,11 @@ pub enum Error {
     },
     #[snafu(display("failed to create cluster resources"))]
     CreateClusterResources {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to delete orphaned resources"))]
     DeleteOrphanedResources {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to resolve the Vector aggregator address"))]
     ResolveVectorAggregatorAddress { source: product_logging::Error },
@@ -149,7 +153,7 @@ pub enum Error {
 
     #[snafu(display("failed to build Metadata"))]
     MetadataBuild {
-        source: stackable_operator::builder::ObjectMetaBuilderError,
+        source: stackable_operator::builder::meta::Error,
     },
 
     #[snafu(display("failed to get required Labels"))]
@@ -189,7 +193,7 @@ pub async fn reconcile(shs: Arc<SparkHistoryServer>, ctx: Arc<Ctx>) -> Result<Ac
     let resolved_product_image = shs
         .spec
         .image
-        .resolve(SPARK_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
+        .resolve(SPARK_IMAGE_BASE_NAME, crate::built_info::PKG_VERSION);
     let s3_log_dir = S3LogDir::resolve(
         Some(&shs.spec.log_file_directory),
         shs.metadata.namespace.clone(),
@@ -310,6 +314,7 @@ pub fn error_policy(_obj: Arc<SparkHistoryServer>, _error: &Error, _ctx: Arc<Ctx
     Action::requeue(*Duration::from_secs(5))
 }
 
+#[allow(clippy::result_large_err)]
 fn build_config_map(
     shs: &SparkHistoryServer,
     config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -371,6 +376,7 @@ fn build_config_map(
         .context(InvalidConfigMapSnafu { name: cm_name })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_stateful_set(
     shs: &SparkHistoryServer,
     resolved_product_image: &ResolvedProductImage,
@@ -512,6 +518,7 @@ fn build_stateful_set(
     })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_service(
     shs: &SparkHistoryServer,
     app_version_label: &str,
@@ -577,6 +584,7 @@ fn build_service(
     })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_history_role_serviceaccount(
     shs: &SparkHistoryServer,
     app_version_label: &str,
@@ -617,6 +625,7 @@ fn build_history_role_serviceaccount(
     Ok((sa, binding))
 }
 
+#[allow(clippy::result_large_err)]
 fn spark_defaults(
     shs: &SparkHistoryServer,
     s3_log_dir: &S3LogDir,
@@ -720,6 +729,7 @@ fn labels<'a, T>(
 /// Return the Spark properties for the cleaner role group (if any).
 /// There should be only one role group with "cleaner=true" and this
 /// group should have a replica count of 0 or 1.
+#[allow(clippy::result_large_err)]
 fn cleaner_config(
     shs: &SparkHistoryServer,
     rolegroup_ref: &RoleGroupRef<SparkHistoryServer>,