+++ rust/operator-binary/src/product_logging.rs
@@ -2,7 +2,7 @@ use std::fmt::Display;
 
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
-    builder::ConfigMapBuilder,
+    builder::configmap::ConfigMapBuilder,
     client::Client,
     k8s_openapi::api::core::v1::ConfigMap,
     kube::Resource,
@@ -21,7 +21,7 @@ use stackable_spark_k8s_crd::constants::{
 pub enum Error {
     #[snafu(display("failed to retrieve the ConfigMap {cm_name}"))]
     ConfigMapNotFound {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::client::Error,
         cm_name: String,
     },
     #[snafu(display("failed to retrieve the entry {entry} for ConfigMap {cm_name}"))]